add authenticate and  jwt token

+++ server/src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from "../lib/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance){
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify();
+  })
+ 
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc'
       }
@@ -18,7 +25,7 @@ export async function memoriesRoutes(app: FastifyInstance){
     })
   })
 
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -27,6 +34,11 @@ export async function memoriesRoutes(app: FastifyInstance){
     const memory = await prisma.memory.findUniqueOrThrow({
       where: { id }
     });
+
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
     return memory
   })
 
@@ -41,13 +53,13 @@ export async function memoriesRoutes(app: FastifyInstance){
 
     const memory = await prisma.memory.create({
       data: {
-        content, coverUrl, isPublic, userId: 'sadasdasdsadasdasd'
+        content, coverUrl, isPublic, userId: request.user.sub
       }
     });
     return memory
   })
 
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
@@ -60,21 +72,38 @@ export async function memoriesRoutes(app: FastifyInstance){
     const { id } = paramsSchema.parse(request.params)
     const { content, coverUrl, isPublic } =  bodySchema.parse(request.body)
 
-   const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
-      data: {
-        content, coverUrl, isPublic
-      }
-    });
+    })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
+        where: { id },
+        data: {
+          content, coverUrl, isPublic
+        }
+      });
     return memory
   })
 
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid()
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id },
+    })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
     await prisma.memory.delete({
       where: { id }
     });
